feat: split out logic
