+++ pyproject.toml
@@ -26,13 +26,6 @@ dependencies = [
     "sodapy>=2.2.0",
 ]
 
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/dataops"]
-
 [dependency-groups]
 dev = [
     "pre-commit>=4.2.0",
@@ -42,3 +35,10 @@ dev = [
 
 [tool.pytest.ini_options]
 pythonpath = ["."]
+
+[tool.uv]
+package = true
+
+[build-system]
+requires = ["uv_build>=0.7.20,<0.8.0"]
+build-backend = "uv_build"
